[WIP] INT-24 QA changes
* zone.secondaries.network_id -> networks (for API parity)
* make.zone.secondaries.notify optional in schema (woops)
* disambiguate a test failure message a bit more
* attempt to make the data resource schema work better (wip)

Issues to figure out:

* zone data resource doesn't work right still. secondaries is always [],
even when they exist. Not sure why this is, since we use the same logic
to populate tfstate in the regular zone resource, and that picks up
secondaries (mostly)
* zone import has same issue as zone data, probably related
* issues with secondary "networks" field:
  * leaving it unset leads to `networks = []` in tfstate. API shows a
    correct result of [0]
  * subsequently setting it to [0] triggers a change in apply. API is
    still correct
  * if we start with `networks = [1]` and then remove the line from
    our tf file, terraform NOOPs on apply. I think we'd prefer that
    to trigger a change and the result in API should be [0]

My suspicion with the [] vs [0] issue is that the SDK is doing some
mangling, but not sure why the `TestAccZone_primary` test passes, as
it's asserting that a `secondaries` entry with no `networks` fields ends
up being set to `[]int{0}`, and it fails if we expect `[]int{}`.

In any event, unset `networks` key should lead to a `[0]` in state, and
subsequent applys should not should detect a change.